refactor: getUserInfos 경로 수정
