+++ src/app/Pcapsule/pcapsuleController.js
@@ -1,7 +1,7 @@
 import { response } from "../../../config/response.js";
 import { status } from "../../../config/responseStatus.js";
 
-import { getUserInfos } from "../User/userService.js";
+import { getUserInfos } from "../User/userProvider.js";
 
 import { createPcs_s, readPcs_s, readDetailPcs_s } from "./pcapsuleService.js";
 import { savePassword_p } from "./pcapsuleProvider.js";
